# pedidos-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de pedidos",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/order": {
            "post": {
                "tags": [
                    "Pedido"
                ],
                "summary": "Crear un pedido",
                "operationId": "e6f6ccd33c4446106ff016845490d367",
                "parameters": [
                    {
                        "name": "products",
                        "in": "query",
                        "description": "Lista de productos",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "delivery_location_id",
                        "in": "query",
                        "description": "Id de la ubicación de entrega",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type_delivery",
                        "in": "query",
                        "description": "Tipo de pago 1 = efectivo, 2 = tarjeta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pedido creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Pedido creado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error al crear el pedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error al crear el pedido"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No autorizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error interno del servidor"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/orders": {
            "get": {
                "summary": "Retrieve orders",
                "description": "Retrieves a list of orders based on certain criteria.",
                "operationId": "2cd7e03021e72d9d799c12dfe179c865",
                "parameters": [
                    {
                        "name": "is_deliver",
                        "in": "query",
                        "description": "Filter orders by deliverer (true/false)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of orders",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "data": [
                                        {
                                            "id": 1,
                                            "type_pay": "Efectivo",
                                            "status_text": "Nuevo"
                                        },
                                        {
                                            "id": 2,
                                            "type_pay": "Tarjeta",
                                            "status_text": "En camino"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "Internal Server Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "get": {
                "summary": "Get order details",
                "description": "Retrieves detailed information about a specific order.",
                "operationId": "693bab3f776ed920d3fc277a4871e819",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "data": {
                                        "id": 1,
                                        "type_pay": "Efectivo",
                                        "status_text": "Nuevo",
                                        "products": "#/components/schemas/products",
                                        "delivery_location": "#/components/schemas/delivery_location",
                                        "user": "#/components/schemas/user",
                                        "deliver": "#/components/schemas/deliver"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "Order not found"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/orders/{id}/status": {
            "put": {
                "summary": "Change order status",
                "description": "Updates the status of a specific order.",
                "operationId": "dea02acba2a89c95d9ef38eba2a37456",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "JSON object containing the new status",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "description": "New status code",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Se ha actualizado el estado del pedido"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "Validation error details"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "No se encontró el pedido"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}